Handle UTC on date when using 'ccc' on dateformat
ccc doesn't seem to recognize UTC so I'm just replacing it with GMT

For me this is only a problem on Travis or other test machines as
locally I get a proper timezone name from it.

I'm not sure if the fix is technically correct but will have to do for
now.